+++ lib/ui/BrowsePanel.dart
@@ -60,6 +60,8 @@ import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
 import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
+import 'dining/DiningLinksPanel.dart';
+
 ///////////////////////////
 // BrowsePanel
 
@@ -646,6 +648,7 @@ class _BrowseEntry extends StatelessWidget {
       case "career_exploration.job_board":   _onTapJobBoard(context); break;
 
       case "dining.dining":                  _onTapDining(context); break;
+      case "dining.dining_links":            _onTapDiningLinks(context); break;
 
       case "directory.user_directory":       _onTapUserDirectory(context); break;
 
@@ -749,8 +752,13 @@ class _BrowseEntry extends StatelessWidget {
   }
 
   static void _onTapDining(BuildContext context) {
-    Analytics().logSelect(target: "Residence Hall Dining");
+    Analytics().logSelect(target: "University Housing Dining");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => Dining2HomePanel()));
+  }
+
+  static void _onTapDiningLinks(BuildContext context) {
+    Analytics().logSelect(target: "Dining Links");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => DiningLinksPanel()));
   }
 
   static void _onTapUserDirectory(BuildContext context) {